Document how to update the image using static delta
The README now has the instruction set on how to build two images wheree
the second is an updated version of the first one. Doing this while
recording the ostree repository so that it can be used to generate an
ostree static delta file that is then used to update from the first
image to the second, allowing to move from a system that does not
support installing securely containerized application to one that does.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>